parsing, updating a student, sending a file functions work

+++ peer-assessment-backend/src/main/java/com/monaco/peer_assessment_backend/dto/StudentDTO.java
@@ -16,7 +16,7 @@ public class StudentDTO extends UserDTO {
     private long studentId;
     private boolean isTemp;
 
-    public StudentDTO(long id, String firstName, String lastName, Set<String> roles, boolean isTemp, int studentId) {
+    public StudentDTO(long id, String firstName, String lastName, Set<String> roles, boolean isTemp, long studentId) {
         super(id, null, firstName, lastName, " ", roles);
         this.isTemp = isTemp;
         this.studentId = studentId;
